+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 
 - Command: `dev-proxy-toolkit.config-open` - Open configuration file
+- Command: `dev-proxy-toolkit.config-new` - Create configuration file
 
 ### Changed: