+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Command: `dev-proxy-toolkit.config-new` - Create new configuration file
 - Command: `dev-proxy-toolkit.restart` - Restart Dev Proxy
 
+### Changed:
+
+- Snippets: Updated all snippets to use `v0.27.0` schema
+
 ## [0.20.0] - 2025-04-01
 
 ### Added: